pyproject.toml
[tool.poetry]
name = "langchain-vector"
version = "0.1.0"
description = ""
authors = ["daichi_yamaoka <tnychor06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.7"
langchain-community = "^0.3.7"
dashvector = "^1.0.18"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
